// src/generated/types.ts
export type UserDbObject = {
  _id: ObjectID,
  name: string,
};